requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
email-validator==2.1.0.post1
